eric-cnam-onboarding-server/src/main/java/com/ericsson/oss/management/onboarding/presentation/controllers/HealthController.java
/*******************************************************************************
 * COPYRIGHT Ericsson 2023
 *
 *
 *
 * The copyright to the computer program(s) herein is the property of
 *
 * Ericsson Inc. The programs may be used and/or copied only with written
 *
 * permission from Ericsson Inc. or in accordance with the terms and
 *
 * conditions stipulated in the agreement/contract under which the
 *
 * program(s) have been supplied.
 ******************************************************************************/

package com.ericsson.oss.management.onboarding.presentation.controllers;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.ericsson.oss.management.onboarding.api.HealthApi;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;

@Slf4j
@RestController
@RequestMapping("/cnonb/v1")
@RequiredArgsConstructor
public class HealthController implements HealthApi {

    @Override
    public ResponseEntity<Void> healthGet() {
        log.info("Received a Get request to get health of onboarding service");
        return new ResponseEntity<>(HttpStatus.OK);
    }

}
