eric-cnam-onboarding-server/src/test/java/com/ericsson/oss/management/onboarding/presentation/controllers/HealthControllerTest.java
/*******************************************************************************
 * COPYRIGHT Ericsson 2023
 *
 *
 *
 * The copyright to the computer program(s) herein is the property of
 *
 * Ericsson Inc. The programs may be used and/or copied only with written
 *
 * permission from Ericsson Inc. or in accordance with the terms and
 *
 * conditions stipulated in the agreement/contract under which the
 *
 * program(s) have been supplied.
 ******************************************************************************/

package com.ericsson.oss.management.onboarding.presentation.controllers;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.HttpStatus;

@SpringBootTest(classes = HealthController.class)
class HealthControllerTest {

    @Autowired
    private HealthController controller;

    @Test
    void shouldReturnResponseWhenGetHealth() {
        //Test method
        HttpStatus result = controller.healthGet()
                .getStatusCode();

        //Verify
        assertThat(result).isEqualTo(HttpStatus.OK);
    }
}
